backend/requirements.txt
flask 
flask-cors 
joblib 
python-dotenv 
pandas
pickle
groq